# docs/api/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: hello@goatflow.io
    name: GoatFlow Support
    url: https://goatflow.io/support
  description: GoatFlow Ticket System REST API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://goatflow.io/terms/
  title: GoatFlow API
  version: "1.0"
paths:
...
  /webhooks:
    get:
      description: List outbound webhooks (admin only)
      parameters:
      - description: Only active (true) or inactive (false) webhooks
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 'success, data: []Webhook'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid filter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Create an outbound webhook; deliveries are signed with X-Webhook-Signature
        (sha256=HMAC of the body) when a secret is set
      parameters:
      - description: name, url, events (required); secret, headers, retry_count, timeout_seconds,
          is_active
        in: body
        name: webhook
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 'success, data: Webhook'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Name already used
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create webhook
      tags:
      - Webhooks
  /webhooks/{id}:
    delete:
      description: Deletes the webhook and its delivery log
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete webhook
      tags:
      - Webhooks
    get:
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'success, data: Webhook'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get webhook
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: 'Partial update: only fields present change; an empty secret removes
        it'
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to change
        in: body
        name: webhook
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'success, data: Webhook'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Name already used
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update webhook
      tags:
      - Webhooks
...
  /webhooks/{id}/test:
    post:
      description: Sends a webhook.test event now (one attempt) and returns the recorded
        delivery
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'success, data: WebhookDelivery'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Send test delivery
      tags:
      - Webhooks
...
  /webhooks/events:
    get:
      description: Events a webhook can subscribe to
      produces:
      - application/json
      responses:
        "200":
          description: 'success, data: [{event, description}]'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List webhook events
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: API token (Bearer gf_xxx) or JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"